Refactor and enhance functionality in trainscanner2 modules
- Updated PathItem class to use float coordinates for improved precision.
- Introduced BlurMask3 class for enhanced frame comparison, allowing detection of slower trains.
- Modified analyze_iter to utilize the new BlurMask3 for better motion detection.
- Improved logging in MotionDetector to capture detailed frame processing information.
- Updated remove_path methods in MultiViewWindow and MultiViewManager to include reason for path removal.
- Adjusted rendering logic to prevent immediate closure of low-quality windows based on user feedback.
- Enhanced ImageStripsWidget to utilize shapes for width calculation, improving performance.

+++ trainscanner2/__init__.py
@@ -7,7 +7,7 @@ from trainscanner.image import standardize
 
 @dataclass
 class PathItem:
-    xy: tuple[int, int]
+    xy: tuple[float, float]
     value: list
